+++ src/main/java/com/agenticcp/core/domain/cloud/adapter/outbound/aws/account/AwsAccountCredentialManagementAdapter.java
@@ -8,7 +8,7 @@ import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import com.agenticcp.core.domain.cloud.port.model.account.CloudSessionCredential;
 import com.agenticcp.core.domain.cloud.port.outbound.account.AccountCredentialManagementPort;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
-import com.agenticcp.core.domain.cloud.service.SessionCacheService;
+import com.agenticcp.core.domain.cloud.service.account.SessionCacheService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
