+++ src/main/java/com/agenticcp/core/domain/cloud/service/account/CloudAccountUseCaseService.java
@@ -1,4 +1,4 @@
-package com.agenticcp.core.domain.cloud.service;
+package com.agenticcp.core.domain.cloud.service.account;
 
 import com.agenticcp.core.common.context.TenantContextHolder;
 import com.agenticcp.core.common.enums.CommonErrorCode;
