+++ src/test/java/com/agenticcp/core/domain/cloud/adapter/outbound/aws/account/AwsAccountCredentialManagementAdapterTest.java
@@ -7,7 +7,7 @@ import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import com.agenticcp.core.domain.cloud.port.model.account.CloudSessionCredential;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
-import com.agenticcp.core.domain.cloud.service.SessionCacheService;
+import com.agenticcp.core.domain.cloud.service.account.SessionCacheService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
