+++ src/test/java/com/agenticcp/core/domain/cloud/service/CloudAccountDomainServiceTest.java
@@ -7,6 +7,7 @@ import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import com.agenticcp.core.domain.cloud.enums.AccountStatus;
 import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import com.agenticcp.core.domain.cloud.repository.CloudAccountRepository;
+import com.agenticcp.core.domain.cloud.service.account.CloudAccountDomainService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
