+++ src/test/java/com/agenticcp/core/domain/cloud/service/CloudAccountUseCaseServiceTest.java
@@ -11,6 +11,7 @@ import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import com.agenticcp.core.domain.cloud.enums.AccountStatus;
 import com.agenticcp.core.domain.cloud.exception.CloudErrorCode;
 import com.agenticcp.core.domain.cloud.mapper.CredentialCommandMapper;
+import com.agenticcp.core.domain.cloud.service.account.CloudAccountUseCaseService;
 import com.agenticcp.core.domain.cloud.port.outbound.account.AccountSyncPort;
 import com.agenticcp.core.domain.cloud.port.outbound.account.AccountValidationPort;
 import com.agenticcp.core.domain.cloud.port.outbound.AuditEventPort;
